Record videos locally
Instead of trying to stream recordings, or send recordings over the
wire, I decided it's simpler if we just let each client record its own
video. This will make it more likely that all clients will be able to
play the videos they've recorded. Also, we don't need to send photos and
such over the wire.