Switch auto-record strategy to floating window
Instead of drawing on a delayed canvas, which turned out to have bad
performance and terrible cross-browser support, I'm now changing to
using a floating window. We start a recording every second, keep it
around for 2s and then discard it, unless we've deteced a auto-record
trigger, in which case we keep it around for another 2s, saving it at
the end.

So far only the broadcaster side has been updated, but I'll get to the
watch page next.