Use sliding window auto recorder on watch page as well
This commit connects the broadcaster and the watcher again by making the
AutoRecorder work in "passive" mode on the watch side of things.

To continue making videos look nice in chrome when a mobile is used to
transmit, I had to use a fallback video. This time I'm ditching the
DelayedVideo and using a simpler FallbackVideo. Only enabled in chrome,
since other browsers seem to do the right thing.